+++ examples/lib/stories/sprites/spritesheet.dart
@@ -13,8 +13,17 @@ class SpritesheetGame extends FlameGame {
 
     final vampireAnimation =
         spriteSheet.createAnimation(row: 0, stepTime: 0.1, to: 7);
+
     final ghostAnimation =
         spriteSheet.createAnimation(row: 1, stepTime: 0.1, to: 7);
+
+    final ghostAnimationVariableStepTimes =
+        spriteSheet.createAnimationWithVariableStepTimes(
+      row: 1,
+      to: 7,
+      stepTimes: [0.1, 0.1, 0.3, 0.3, 0.5, 0.3, 0.1],
+    );
+
     final spriteSize = Vector2(80.0, 90.0);
 
     final vampireComponent = SpriteAnimationComponent(
@@ -29,8 +38,15 @@ class SpritesheetGame extends FlameGame {
       size: spriteSize,
     );
 
+    final ghostAnimationVariableStepTimesComponent = SpriteAnimationComponent(
+      animation: ghostAnimationVariableStepTimes,
+      position: Vector2(150, 340),
+      size: spriteSize,
+    );
+
     add(vampireComponent);
     add(ghostComponent);
+    add(ghostAnimationVariableStepTimesComponent);
 
     // Some plain sprites
     final vampireSpriteComponent = SpriteComponent(
@@ -45,7 +61,14 @@ class SpritesheetGame extends FlameGame {
       position: Vector2(50, 220),
     );
 
+    final ghostVariableSpriteComponent = SpriteComponent(
+      sprite: spriteSheet.getSprite(1, 0),
+      size: spriteSize,
+      position: Vector2(50, 340),
+    );
+
     add(vampireSpriteComponent);
     add(ghostSpriteComponent);
+    add(ghostVariableSpriteComponent);
   }
 }