+++ packages/flame/lib/src/components/nine_tile_box_component.dart
@@ -2,6 +2,7 @@ import 'dart:ui';
 
 import 'package:meta/meta.dart';
 
+import '../../components.dart';
 import '../extensions/vector2.dart';
 import '../nine_tile_box.dart';
 import 'position_component.dart';
@@ -19,8 +20,18 @@ class NineTileBoxComponent extends PositionComponent {
     this.nineTileBox, {
     Vector2? position,
     Vector2? size,
+    Vector2? scale,
+    double? angle,
+    Anchor? anchor,
     int? priority,
-  }) : super(position: position, size: size, priority: priority);
+  }) : super(
+          position: position,
+          size: size,
+          scale: scale,
+          angle: angle,
+          anchor: anchor,
+          priority: priority,
+        );
 
   @mustCallSuper
   @override